Expose v2 mod and rename borrowed to model.
