+++ src/lib.rs
@@ -21,6 +21,7 @@ pub mod model;
 pub mod error;
 
 pub mod v1;
+pub mod v2;
 
 /// Parses a version 1 header of HAProxy's proxy protocol.
 /// Supports TCP with IPv4 and IPv6 addresses, as well as UNKNOWN address information.
